Split OMGTextInput into multiple components
