+++ src/components/widgets/index.js
@@ -1,21 +1,25 @@
 import OMGAppBar from './omg-app-bar'
+import OMGBackground from './omg-background'
+import OMGBox from './omg-box'
 import OMGButton from './omg-button'
 import OMGCard from './omg-card'
+import OMGIcon from './omg-icon'
+import OMGPasswordTextInput from './omg-password-text-input'
+import OMGRadioButton from './omg-radio-button'
+import OMGText from './omg-text'
 import OMGTextInput from './omg-text-input'
 import OMGTitle from './omg-title'
-import OMGText from './omg-text'
-import OMGIcon from './omg-icon'
-import OMGBackground from './omg-background'
-import OMGRadioButton from './omg-radio-button'
 
 export {
   OMGAppBar,
+  OMGBackground,
+  OMGBox,
   OMGButton,
   OMGCard,
+  OMGIcon,
+  OMGPasswordTextInput,
+  OMGRadioButton,
   OMGTextInput,
   OMGTitle,
-  OMGText,
-  OMGIcon,
-  OMGBackground,
-  OMGRadioButton
+  OMGText
 }
