+++ src/components/widgets/omg-button/index.js
@@ -1,0 +1,6 @@
+import React from 'react'
+import { TouchableOpacity } from 'react-native'
+
+const OMGButton = props => {}
+
+export default OMGButton
